Testing to calculate 100 factorial and change signature to calculate method for return long type

+++ Factorial/FactorialHandler.cs
@@ -2,7 +2,7 @@ namespace Factorial
 {
     public static class FactorialHandler
     {
-        public static int Calculate(int number)
+        public static long Calculate(int number)
         {
             return number <= 1 ? 1 : number * Calculate(number - 1);
         }
